FEA Add djangorestframework dependency and update settings
* Add djangorestframework to base requirements
* Update dev and prod settings to use correct relative imports for common settings
* Load environment variables using python-dotenv in dev and prod settings

The commit adds the djangorestframework package to the base requirements file, which is a popular library for building RESTful APIs in Django. It also updates the dev.py and prod.py settings files to use the correct relative imports for the common settings and to load environment variables using the python-dotenv package. These changes lay the foundation for integrating RESTful API functionality into the Django project.