[FEA] Add AWS CDK infrastructure for Django ECS deployment
* Set up a new AWS CDK project with TypeScript
* Define VPC, ECS cluster, and Fargate service resources
* Configure environment variables and secrets for the Django container
* Specify the Django container image and port mappings
* Output the load balancer DNS name for accessing the deployed application

This commit adds the AWS CDK infrastructure code for deploying a Django application on ECS Fargate. It sets up the necessary resources such as VPC, ECS cluster, and Fargate service. The Django container is configured with environment variables and secrets, and the container image and port mappings are specified. The load balancer DNS name is outputted for accessing the deployed application.