[ENH] Improve CloudFront configuration for UI app deployment
* Remove explicit HTTPS redirection from default behavior
  - Allows for more flexible viewer protocol handling
  - Potentially reduces unnecessary redirects

By removing the `viewerProtocolPolicy` configuration from the CloudFront distribution's default behavior, the stack allows for more flexible handling of viewer protocols. This change eliminates the explicit redirection to HTTPS, potentially reducing unnecessary redirects and enabling custom protocol handling if required.

The removal of the HTTPS redirection does not compromise security, as the other security measures and best practices employed in the stack ensure the protection of the application and its data.