[ENH] Enable dark theme by default
* Update default value of `darkTheme` to `true` in `default.vue` layout
  - Provides a dark theme experience for users upon initial load
  - Aligns with modern design trends and user preferences

This change sets the default theme of the application to dark mode. By initializing the `darkTheme` data property to `true`, the UI will render with a dark color scheme when the user first visits the site or refreshes the page. This enhancement aims to provide a visually appealing and comfortable experience for users who prefer darker interfaces.