FEA Update project name and references
* Rename project from "trellis-law" to "number-2-words" across the codebase
  - Update project name in README files, comments, and configurations
  - Modify asset names and paths to reflect the new project name
* Update Docker image tag from "lguibr/django-trellis-example" to "lguibr/django-rest"
* Change S3 bucket name for UI deployment to "bucket-number-2-words-ui-adhoc"
* Update GitHub repository links to point to "lguibr/number-2-words"

This commit renames the project from "trellis-law" to "number-2-words" to better reflect its purpose. It updates all references to the old name across the codebase, including in README files, comments, configurations, asset names, and paths. The Docker image tag for the backend is also updated, and the S3 bucket name for UI deployment is changed accordingly. Finally, GitHub repository links are modified to point to the new "number-2-words" repository.